requirements.txt
Flask==3.1.1
mysql-connector-python==9.4.0
gunicorn==23.0.0
boto3==1.39.11
python-dotenv==1.1.1
